Added email, password, validator, bcrypt and jwt

+++ routes/author.routes.js
@@ -1,8 +1,11 @@
 const express = require("express");
 const { Book } = require("../models/Book");
 const fs = require("fs");
+const bcrypt = require("bcrypt");
 const multer = require("multer");
 const upload = multer({ dest: "public" });
+const { generateToken } = require("../utils/token");
+const { isAuth } = require("../middlewares/auth.middleware.js");
 
 // Modelos
 const { Author } = require("../models/Author.js");
@@ -58,7 +61,7 @@ router.get("/", async (req, res, next) => {
 router.get("/:id", async (req, res, next) => {
   try {
     const id = req.params.id;
-    const author = await Author.findById(id);
+    const author = await Author.findById(id).select("+password");
 
     if (author) {
       const temporalAuthor = author.toObject();
@@ -92,6 +95,43 @@ router.get("/name/:name", async (req, res, next) => {
   }
 });
 
+// LOGIN DE AUTORES
+router.post("/login", async (req, res, next) => {
+  try {
+    // const email = req.body.email;
+    // const password = req.body.password;
+    const { email, password } = req.body;
+
+    if (!email || !password) {
+      return res.status(400).json({ error: "Se deben especificar los campos email y password" });
+    }
+
+    const author = await Author.findOne({ email }).select("+password");
+    if (!author) {
+      // return res.status(404).json({ error: "No existe un usuario con ese email" });
+      // Por seguridad mejor no indicar qué usuarios no existen
+      return res.status(401).json({ error: "Email y/o contraseña incorrectos" });
+    }
+
+    // Comprueba la pass
+    const match = await bcrypt.compare(password, author.password);
+    if (match) {
+      // Quitamos password de la respuesta
+      const authorWithoutPass = author.toObject();
+      delete authorWithoutPass.password;
+
+      // Generamos token JWT
+      const jwtToken = generateToken(author._id, author.email);
+
+      return res.status(200).json({ token: jwtToken });
+    } else {
+      return res.status(401).json({ error: "Email y/o contraseña incorrectos" });
+    }
+  } catch (error) {
+    next(error);
+  }
+});
+
 // CRUD: CREATE
 router.post("/", async (req, res, next) => {
   try {
@@ -105,9 +145,14 @@ router.post("/", async (req, res, next) => {
 });
 
 // CRUD: DELETE
-router.delete("/:id", async (req, res, next) => {
+router.delete("/:id", isAuth, async (req, res, next) => {
   try {
     const id = req.params.id;
+
+    if (req.author.id !== id && req.author.email !== "admin@gmail.com") {
+      return res.status(401).json({ error: "No tienes autorización para realizar esta operación" });
+    }
+
     const authorDeleted = await Author.findByIdAndDelete(id);
     if (authorDeleted) {
       res.json(authorDeleted);
@@ -120,12 +165,22 @@ router.delete("/:id", async (req, res, next) => {
 });
 
 // CRUD: UPDATE
-router.put("/:id", async (req, res, next) => {
+router.put("/:id", isAuth, async (req, res, next) => {
   try {
     const id = req.params.id;
-    const authorUpdated = await Author.findByIdAndUpdate(id, req.body, { new: true, runValidators: true });
-    if (authorUpdated) {
-      res.json(authorUpdated);
+
+    if (req.author.id !== id && req.author.email !== "admin@gmail.com") {
+      return res.status(401).json({ error: "No tienes autorización para realizar esta operación" });
+    }
+
+    const authorToUpdate = await Author.findById(id);
+    if (authorToUpdate) {
+      Object.assign(authorToUpdate, req.body);
+      await authorToUpdate.save();
+      // Quitamos pass de la respuesta
+      const authorToSend = authorToUpdate.toObject();
+      delete authorToSend.password;
+      res.json(authorToSend);
     } else {
       res.status(404).json({});
     }
@@ -133,6 +188,7 @@ router.put("/:id", async (req, res, next) => {
     next(error);
   }
 });
+
 router.post("/logo-upload", upload.single("logo"), async (req, res, next) => {
   try {
     // Renombrado de la imagen
